Add files via upload
The main idea here is to classify the sms text into Spams/Ham based on the given classes by applying perhaps all the classification algorithms viz KNN,DecisionTree,RandomForest,SVM etc available as well as ensemble methods to try and get the best model for the above business problem. But before going for model building, majority of the effort was consumed in pre-processing steps which primarily involved removal of emails, phone numbers, other numbers etc, Tokenization, Stop word removal, Punctuations removal, Stemming etc. It was after this that we could proceed for a model building phase.